refactor : 소스 리팩토링
* 각 클래스의 접근 제한자 수정
* 자동차이름에 대한 테스트 분리
